Working on an arrow overlay. - Unified #draw(PGraphics) and #drawOn(PImage) in MapVizualization, adapting Applet in package ui (better name?) to use PGraphics instead of PImage as background buffer. - Added classes Layers and ArrowOverlay, both extending MapVisualization. - Not happy with current inheritance of Layers and inner Arrow class, there should be a common Drawable class that is both drawable and can has children. And then model the whole drawables inherit from that. - Also not happy with Locations, they should offer both normalized and pixel coordinates. If locations belong to one map only, it is save to cache pixel coordinates there.

+++ Students/erni/master/meander/trunk/src/ch/deif/meander/Location.java
@@ -4,14 +4,14 @@ import ch.akuhn.hapax.corpus.Document;
 
 public class Location {
 
-    public double x, y;
-    public double height;
+    public float x, y;
+    public float height;
     public Document document;
 
     public Location(double xNormed, double yNormed, double height) {
-        this.x = xNormed;
-        this.y = yNormed;
-        this.height = height;
+        this.x = (float) xNormed;
+        this.y = (float) yNormed;
+        this.height = (float) height;
     }
 
 }
